Refactor Instrument::from_json to cast tick_size as size_t for consistency

+++ server_app/instrument/instrument.cpp
@@ -20,7 +20,7 @@ Instrument Instrument::from_json(Json& data)
         data["symbol"],
         data["exchange_symbol"],
         (size_t)data["lot_size"],
-        (double)data["tick_size"]
+        (size_t)data["tick_size"]
     };
 };
